Fix crash on driversim exit
The plugin wasn't being deconstructed properly, causing some crashes on
the exit code. That works but isn't very graceful. To fix this, we check
if we have a shutdown flag set to stop the main publish thread. We also
unregister our auto-state subscriber.

I know that the boolean flag isn't the best for threading, but it
works okay for this case since we only ever set the `stop_pub` flag once
during the entire life of the program.