Use a simpler algorithm to group beacons by type
Not the one Kevin recommended, but I think it's even simpler.
It maps each beacon type to a struct with:
- Field beacons of that type
- Distances of the detected beacons to the field beacons
First, the field beacons are copied into the map.
Then, each detected beacon has its distances calculated and pushed to
the vector for its type.
It still uses the Hungarian algorithm to match detections to actual
beacons, and adds up the distances.